# perf: cached_property on early reject is slow