`PROFILE=0 python -O test/external/external_benchmark_schedule.py` best out of 5

Master
```
***** model tensor in     17.41 ms
***** model schedule in    7.25 ms
***** model opts(53) in   29.08 ms
***** model lower in       6.35 ms
***** model rewrite in   477.46 ms
***** model linearize in 127.28 ms
14964
all 668.97 ms
```

Branch
```
***** model tensor in     17.48 ms
***** model schedule in    7.28 ms
***** model opts(53) in   28.40 ms
***** model lower in       6.30 ms
***** model rewrite in   454.29 ms
***** model linearize in 121.87 ms
14964
all 639.66 ms
```