@geohot 2024-08-29
Hmm, interesting. Why is cached_property so terrible?