@geohot 2024-08-29
Oh this should just be refactored to precompute it