@geohot 2024-08-30
The best part is no part. #6327